Day10_main.dart
This is a simple Flutter app that functions as a counter. When you press the button with the plus_one icon, the counter value increases by one. The app uses a stateful widget, CounterApp, to keep track of and display the count, updating the UI each time the button is pressed. The main screen consists of an app bar with a title and a central area with two text widgets displaying information and the count value. An elevated button triggers the CountNumber function to update the count through setState(), ensuring a smooth, reactive UI update.